+++ day17/src/grid.rs
@@ -1,7 +1,7 @@
 use crate::pos::{Dir, Pos};
 
 pub struct Grid {
-    pub data: Vec<Vec<usize>>,
+    pub data: Vec<Vec<u8>>,
 }
 impl From<Vec<String>> for Grid {
     fn from(mat: Vec<String>) -> Self {
@@ -10,7 +10,7 @@ impl From<Vec<String>> for Grid {
         for line in mat.iter() {
             let mut data_line = Vec::new();
             for ch in line.chars() {
-                data_line.push(ch.to_string().parse::<usize>().unwrap())
+                data_line.push(ch.to_string().parse::<u8>().unwrap())
             }
             data.push(data_line);
         }
@@ -26,14 +26,14 @@ impl Grid {
         self.data.len()
     }
 
-    fn get(&self, pos: Pos) -> Option<&usize> {
+    fn get(&self, pos: Pos) -> Option<&u8> {
         self.data
             .get(pos.1 as usize)
             .map(|line| line.get(pos.0 as usize))
             .flatten()
     }
 
-    pub fn neighbours(&self, pos: Pos) -> impl Iterator<Item = (Dir, Pos, &usize)> + '_ {
+    pub fn neighbours(&self, pos: Pos) -> impl Iterator<Item = (Dir, Pos, &u8)> + '_ {
         [Dir::NORTH, Dir::EAST, Dir::SOUTH, Dir::WEST]
             .iter()
             .filter_map(move |&dir| Some((dir, pos + dir, self.get(pos + dir)?)))
